Many

+++ lib/Timer/Timer.h
@@ -11,7 +11,7 @@
 #define TMR64(T, CODE)            \
   {                               \
     static uint64_t tmr##T = 0;   \
-    if ((millis() - tmr##T) >= T) \
+    if ((uint64_t)(millis() - tmr##T) >= T) \
     {                             \
       tmr##T = millis();          \
       CODE                        \
@@ -30,7 +30,7 @@
 #define TMR32(T, CODE)                        \
   {                                           \
     static uint32_t tmr##T = 0;               \
-    if ((((uint32_t)millis()) - tmr##T) >= T) \
+    if ((uint32_t)(((uint32_t)millis()) - tmr##T) >= T) \
     {                                         \
       tmr##T = millis();                      \
       CODE                                    \
@@ -49,7 +49,7 @@
 #define TMR16(T, CODE)                        \
   {                                           \
     static uint16_t tmr##T = 0;               \
-    if ((((uint16_t)millis()) - tmr##T) >= T) \
+    if ((uint16_t)(((uint16_t)millis()) - tmr##T) >= T) \
     {                                         \
       tmr##T = millis();                      \
       CODE                                    \
@@ -68,7 +68,7 @@
 #define TMR8(T, CODE)                        \
   {                                          \
     static uint8_t tmr##T = 0;               \
-    if ((((uint8_t)millis()) - tmr##T) >= T) \
+    if ((uint8_t)(((uint8_t)millis()) - tmr##T) >= T) \
     {                                        \
       tmr##T = millis();                     \
       CODE                                   \
@@ -87,7 +87,7 @@
 #define TMR8_S(T, CODE)                      \
   {                                          \
     static uint8_t tmr##T = 0;               \
-    if ((((uint8_t)millis()) - tmr##T) >= T) \
+    if ((uint8_t)(((uint8_t)millis()) - tmr##T) >= T) \
     {                                        \
       tmr##T += T;                           \
       CODE                                   \
@@ -106,7 +106,7 @@
 #define TMR16_S(T, CODE)                      \
   {                                           \
     static uint16_t tmr##T = 0;               \
-    if ((((uint16_t)millis()) - tmr##T) >= T) \
+    if ((uint16_t)(((uint16_t)millis()) - tmr##T) >= T) \
     {                                         \
       tmr##T += T;                            \
       CODE                                    \
@@ -125,7 +125,7 @@
 #define TMR32_S(T, CODE)                      \
   {                                           \
     static uint32_t tmr##T = 0;               \
-    if ((((uint32_t)millis()) - tmr##T) >= T) \
+    if ((uint32_t)(((uint32_t)millis()) - tmr##T) >= T) \
     {                                         \
       tmr##T += T;                            \
       CODE                                    \
@@ -144,7 +144,7 @@
 #define TMR64_S(T, CODE)          \
   {                               \
     static uint64_t tmr##T = 0;   \
-    if ((millis() - tmr##T) >= T) \
+    if ((uint64_t)(millis() - tmr##T) >= T) \
     {                             \
       tmr##T += T;                \
       CODE                        \
